feat: :sparkles: add connection to movie list endpoint

+++ lib/models/movies_model.dart
@@ -139,7 +139,7 @@ class MovieDetailsModel {
     releaseDate = json['release_date'];
     title = json['title'];
     video = json['video'];
-    voteAverage = json['vote_average'];
+    voteAverage = json['vote_average'].toDouble();
     voteCount = json['vote_count'];
   }
